Refactor module name to format
