+++ src/vecsync/chat/format.py
